Add setgid wrapper
This commit adds a gid C wrapper. The idea:
The actual script processing input events is renamed to
libinput-gestures.py. Additionally, a C programm is added which is
called libinput-gestures (compiled from libinput-gestures.c). During
installation, the wrapper is compiled. The owning group of the
executable is set to 'input', additionally, the setgid bit is set on the
executable. This allows to run the gesture wrapper without the current
user having to be in the 'input' group.